Audit cleanup batch 3: dedup git rev-parse, validate_wiki_id, password resolution; skip duplicate resolve_instance
Four small items off #430.

1. playbooks/_pull_latest_playbooks.yml — drop the redundant
   `git rev-parse` call inside the BUILD_COMMIT/BUILD_DATE shell
   step. The "Get new commit hash" task already captured the value
   we'd be re-shelling for; reuse it via copy: content. Replace the
   compound shell with two `copy:` tasks driven from already-
   captured registers (commit + new git-log query for the build
   date). Down from 3 git rev-parse invocations to 2.

2. roles/common/module_utils/canasta_validate.py — extract the
   verbatim `validate_wiki_id` and `RESERVED_WIKI_IDS` from
   canasta_wikis_yaml.py and canasta_farmsettings.py into a single
   module_utils file. Both modules now import via
   `from ansible.module_utils.canasta_validate import …` (the
   role-local `module_utils/` directory; Ansible adds it to its
   loader search path automatically). Unit tests stash the same
   module under `ansible.module_utils.canasta_validate` in
   `sys.modules` so the in-module import resolves under direct
   `pytest` invocation too. Drops the "Keep both copies in sync"
   comment that wasn't keeping anything in sync.

3. roles/create/tasks/_resolve_password.yml — extract the
   "use-provided-or-generate" pattern that was duplicated three
   times for `_rootdbpass`, `_wikidbpass`, `_admin_password`. Helper
   sets `_resolved_password`; the caller renames into the fact it
   wants. Wiki DB password keeps its `wikidbuser`-conditional
   wrapper but reuses the helper for the actual generation step.

4. instance_lifecycle/start.yml + restart.yml — gate the
   `resolve_instance.yml` include on `instance_path is not defined`.
   Top-level `canasta start` / `canasta restart` still trigger it;
   nested callers (config set/unset, add, backup restore, devmode
   enable/disable, upgrade) skip it because they've already
   resolved the instance once at their own entry point.

Touches 7 files net, 77 insertions / 81 deletions. Test suite up
from 654 to 654 (no test count change; conftest wiring updated to
make the role-local module_utils visible to direct pytest imports).
