+++ roles/create/tasks/_resolve_password.yml
@@ -1,0 +1,32 @@
+---
+# Resolve a password the user may have supplied via flag, falling back
+# to a generated one. Sets `_resolved_password`, which the caller
+# typically renames into the fact it actually wants
+# (`_rootdbpass`, `_wikidbpass`, `_admin_password`).
+#
+# Inputs:
+#   _rp_input        - the value the operator passed (may be '' or
+#                      undefined; both treated as "not provided")
+#   _rp_report_msg   - optional message printed after generation,
+#                      e.g. "Generated root database password (will be
+#                      saved to .env)"
+
+- name: Use the operator-provided password
+  ansible.builtin.set_fact:
+    _resolved_password: "{{ _rp_input }}"
+  when: (_rp_input | default('')) != ''
+
+- name: Generate password when none provided
+  when: (_rp_input | default('')) == ''
+  block:
+    - name: Run password generator
+      ansible.builtin.include_tasks: "{{ canasta_root }}/roles/common/tasks/generate_password.yml"
+
+    - name: Capture generated password
+      ansible.builtin.set_fact:
+        _resolved_password: "{{ _generated_password }}"
+
+    - name: Report generated password
+      ansible.builtin.debug:
+        msg: "{{ _rp_report_msg }}"
+      when: (_rp_report_msg | default('')) != ''
